fix: minimap drag — coerceIn type mismatch (Float vs Int), convert deltaLines*lineHeightPx to Int before coerceIn

+++ android/app/src/main/java/com/codespace/ide/editor/CodeEditor.kt
@@ -1142,7 +1142,7 @@ fun CodeEditor(
                                     change.consume()
                                     coroutineScope.launch {
                                         val deltaLines = (dragAmount.y / miniLineHeightPx).toInt()
-                                        val newScroll = (vScroll.value + deltaLines * lineHeightPx)
+                                        val newScroll = (vScroll.value + (deltaLines * lineHeightPx).toInt())
                                             .coerceIn(0, vScroll.maxValue)
                                         vScroll.scrollTo(newScroll)
                                     }
